pyproject.toml
[tool.poetry]
name = "outage-manager"
version = "0.1.0"
description = "Electricity outage status viewer via telegram bot"
authors = ["Ivan Chukhran <ivan.chuhran@gmail.com>"]
readme = "README.md"
packages = [{include = "outage_manager"}]

[tool.poetry.dependencies]
python = "3.11.8"
aiogram = "^3.10.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
lxml = "^5.2.2"
asyncio = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
